lsp/flycheck: give flymake a try
It's well supported in emacs 26.  Plus, I went through a lot of
trouble to enable multiple backends for flycheck and lsp, but working
in python was way too slow.  Just depend on the python language server
to provide feedback through flymake.

The language server can be configured with a ton of linters.  If it
doesn't support something, look into adding a plugin for it.  That
way, multiple python processes aren't constantly getting invoked,
which I think may explain some of the slowdown in my previous setup.